Refactor dependencies consumption
Extract dependencies definition from defs.bzl. This is conform to the
deploying rules guide: [1] and also a preparation to consume the proto,
python and java rules from Starlak.

Add a deprecation message in closure_repositories macro, when it is
still consumed form defs.bzl file. Consider to remove this deprecation
message when a new release is cut.

[1] https://docs.bazel.build/versions/master/skylark/deploying.html